% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_p.R
\name{adjust_p_bonferroni}
\alias{adjust_p_bonferroni}
\alias{adjust_p_parametric}
\alias{adjust_p_simes}
\title{Calculate adjusted p-values}
\usage{
adjust_p_bonferroni(p, hypotheses)

adjust_p_parametric(p, hypotheses, test_corr = NULL)

adjust_p_simes(p, hypotheses)
}
\arguments{
\item{p}{A numeric vector of p-values (unadjusted, raw), whose values should
be between 0 & 1. The length should match the length of \code{hypotheses}.}

\item{hypotheses}{A numeric vector of hypothesis weights. Must be a vector of
values between 0 & 1 (inclusive). The length should match the length of
\code{p}. The sum of hypothesis weights should not exceed 1.}

\item{test_corr}{(Optional) A numeric matrix of correlations between test
statistics, which is needed to perform parametric tests using
\code{graphicalMCP:::adjust_p_parametric()}. The number of rows and columns of
this correlation matrix should match the length of \code{p}.}
}
\value{
A single adjusted p-value for the intersection hypothesis.
}
\description{
For an intersection hypothesis, an adjusted p-value is the smallest
significance level at which the intersection hypothesis can be rejected.
The intersection hypothesis can be rejected if its adjusted p-value is less
than or equal to \eqn{\alpha}. Currently, there are three test types supported:
\itemize{
\item Bonferroni tests for \code{graphicalMCP:::adjust_p_bonferroni()},
\item Parametric tests \insertCite{xi-2017-unified}{graphicalMCP} for
\code{graphicalMCP:::adjust_p_parametric()},
\itemize{
\item Note that one-sided tests are required for parametric tests.
}
\item Simes tests \insertCite{lu-2016-graphical}{graphicalMCP} for
\code{graphicalMCP:::adjust_p_simes()}.
}
}
\examples{
set.seed(22723)

hypotheses <- c(H1 = 0.5, H2 = 0.25, H3 = 0.25)
p <- c(0.019, 0.025, 0.05)

# Bonferroni test
graphicalMCP:::adjust_p_bonferroni(p, hypotheses)

# Simes test
graphicalMCP:::adjust_p_simes(p, hypotheses)

# Parametric test
# Using the `mvtnorm::GenzBretz` algorithm with `abseps = 1e-6`
corr <- matrix(0.5, nrow = 3, ncol = 3)
diag(corr) <- 1
graphicalMCP:::adjust_p_parametric(p, hypotheses, corr)
}
\references{
\itemize{
\item \insertRef{bretz-2009-graphical}{graphicalMCP}
\item \insertRef{lu-2016-graphical}{graphicalMCP}
\item \insertRef{xi-2017-unified}{graphicalMCP}
}
}
\seealso{
Other graphical tests: 
\code{\link{adjust_weights_parametric}()},
\code{\link{graph_rejection_orderings}()},
\code{\link{graph_test_closure_fast}()},
\code{\link{graph_test_closure}()},
\code{\link{graph_test_shortcut}()},
\code{\link{print.graph_report}()},
\code{\link{test_values_bonferroni}()}
}
\concept{graphical tests}
\keyword{internal}
