% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_graph.R
\name{as_initial_graph.graphMCP}
\alias{as_initial_graph.graphMCP}
\alias{as_initial_graph}
\alias{as_initial_graph.igraph}
\alias{as_gmcp_graph}
\alias{as_gmcp_graph.initial_graph}
\alias{as_igraph}
\alias{as_igraph.initial_graph}
\title{Convert between gMCP and graphicalMCP graph classes}
\usage{
\method{as_initial_graph}{graphMCP}(gmcp_graph)

as_initial_graph(other_graph)

\method{as_initial_graph}{igraph}(igraph)

as_gmcp_graph(graph)

\method{as_gmcp_graph}{initial_graph}(graph)

as_igraph(graph)

\method{as_igraph}{initial_graph}(graph)
}
\arguments{
\item{gmcp_graph}{A gMCP graph}

\item{graph}{An initial graph as returned by \code{\link[=graph_create]{graph_create()}}}
}
\value{
For \code{as_gmcp_graph()}, a gMCP graph object, for \code{as_igraph()}, an
igraph object, and for \code{as_initial_graph()}, a graphicalMCP graph object
}
\description{
Convert between gMCP and graphicalMCP graph classes
}
\examples{
g1 <- random_graph(5)

if (requireNamespace("gMCP", quietly = TRUE)) {
  g2 <- as_gmcp_graph(g1)

  all.equal(g1, as_initial_graph(g2))
}
}
