% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_graph.R
\name{as_gmcp_graph}
\alias{as_gmcp_graph}
\title{Convert between gMCP and graphicalMCP graph classes}
\usage{
as_gmcp_graph(graph)
}
\arguments{
\item{graph}{A graphicalMCP initial_graph}

\item{gmcp_graph}{A gMCP graph}
}
\value{
For \code{as_gmcp_graph()}, a gMCP graph object, and for \code{as_graph()}, a
graphicalMCP graph object
}
\description{
Convert between gMCP and graphicalMCP graph classes
}
\examples{
g1 <- random_graph(5)

if (requireNamespace("gMCP", quietly = TRUE)) g2 <- as_gmcp_graph(g1)

all.equal(g1, as_graph(g2))
}
