% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graphs.R
\name{bonferroni_holm}
\alias{bonferroni_holm}
\title{Example graphs from a variety of papers}
\usage{
bonferroni_holm(n, names = NULL)
}
\arguments{
\item{n}{Number of vertices in the Bonferroni-Holm graph}

\item{names}{Optional names for the hypotheses (Must have length \code{n})}
}
\value{
An S3 object of class \code{mcp_graph}, following the structure of
Bonferroni & Holm: A complete graph with equal weight on each hypothesis
and equal weight on each transition
}
\description{
Example graphs from a variety of papers
}
\examples{
bonferroni_holm(3, names = paste("dose", letters[1:3]))
}
\concept{example-graphs}
