% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_tests.R
\name{bonferroni_test_vals}
\alias{bonferroni_test_vals}
\alias{parametric_test_vals}
\alias{simes_test_vals}
\title{Test hypotheses with the critical values method}
\usage{
bonferroni_test_vals(p, weights, alpha)

parametric_test_vals(p, weights, alpha, corr)

simes_test_vals(p, weights, alpha)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{weights}{A numeric vector of hypothesis weights}

\item{alpha}{A numeric scalar specifying the global significance level for
testing}

\item{corr}{A numeric matrix of correlations between hypotheses' test
statistics}
}
\value{
A data frame with columns specifying the values used to calculate
each hypothesis test
}
\description{
Test hypotheses with the critical values method
}
\examples{
w <- c(H1 = .5, H2 = .5, H3 = 0, H4 = 0)

p <- c(.024, .01, .026, .027)

graphicalMCP:::bonferroni_test_vals(p, w, .05)
graphicalMCP:::parametric_test_vals(p, w, .05, diag(4))
graphicalMCP:::simes_test_vals(p, w, .05)
}
\keyword{internal}
