% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_values.R
\name{test_values_bonferroni}
\alias{test_values_bonferroni}
\alias{test_values_parametric}
\alias{test_values_simes}
\title{Test hypotheses with the adjusted significance method}
\usage{
test_values_bonferroni(p, hypotheses, alpha, intersection = NA)

test_values_parametric(p, hypotheses, alpha, intersection = NA, test_corr)

test_values_simes(p, hypotheses, alpha, intersection = NA)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{hypotheses}{A numeric vector of hypothesis hypotheses}

\item{alpha}{A numeric scalar specifying the global significance level for
testing}

\item{intersection}{A numeric scalar used to track which intersection the
test values are from}

\item{test_corr}{A numeric matrix of correlations between hypotheses' test
statistics}
}
\value{
A data frame with columns specifying the values used to calculate
each hypothesis test
}
\description{
Test hypotheses with the adjusted significance method
}
\examples{
w <- c(H1 = .5, H2 = .5, H3 = 0, H4 = 0)

p <- c(.024, .01, .026, .027)

graphicalMCP:::test_values_bonferroni(p, w, .05)
graphicalMCP:::test_values_parametric(p, w, .05, test_corr = diag(4))
graphicalMCP:::test_values_simes(p, w, .05)
}
\keyword{internal}
