% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_report.R
\name{calc_power_slow}
\alias{calc_power_slow}
\title{Run an MCP power simulation}
\usage{
calc_power_slow(
  graph,
  test_alpha = 0.05,
  test_groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  test_corr = NULL,
  sim_n = 100,
  sim_theta = rep(0, length(graph$hypotheses)),
  sim_corr = diag(length(graph$hypotheses)),
  sim_success = 1:2,
  gamma = NULL
)
}
\arguments{
\item{graph}{An initial graph as returned by \code{create_graph()}, or a gamma
graph as returned by \code{gamma_graph()}}

\item{test_alpha}{A numeric scalar specifying the global level to test at}

\item{test_groups}{A list of numeric vectors specifying hypotheses to test
together}

\item{test_types}{A character vector of tests to apply to the given groups}

\item{test_corr}{Optional if no \code{test_types} are parametric. A numeric matrix
of correlations between hypotheses' test statistics}

\item{sim_n}{An integer scalar specifying how many simulations to run}

\item{sim_theta}{A numeric vector of mean values to use when simulating
p-values. Exactly one mean per hypothesis is needed, and p-values will be
sampled from the multivariate normal distribution}

\item{sim_corr}{A numeric matrix of correlations between hypotheses used to
sample from the multivariate normal distribution to generate p-values}

\item{sim_success}{A numeric vector indicating which hypotheses must be
rejected to consider an experiment a success. It can range from a single
hypothesis to all hypotheses in a graph}

\item{gamma}{Optional if \code{graph} is an initial graph. A numeric gamma matrix
passed on to \code{gamma_graph()}}
}
\value{
A list with five elements
\itemize{
\item power_local - rejection proportion for each hypothesis individually
\item power_expected - average number of hypotheses rejected in a single
simulation
\item power_at_least_1 - proportion of simulations which reject any hypothesis
\item power_all - proportion of simulations which reject all hypotheses
\item power_success - proportion of simulations which reject any of the
hypotheses specified in \code{sim_success}
}
}
\description{
Run an MCP power simulation
}
\examples{
par_gate <- simple_successive_1()

# the default is to test all hypotheses with: Bonferroni testing at alpha
# level .05, 0 mean under the alternative, and 0 correlation between
# hypotheses under the alternative
# the default of 100 simulations will usually need to be increased
calculate_power_vms(par_gate, sim_n = 1e5)

# but any test group/type combination that works for `test_graph()` can be
# used
calculate_power_vms(
  par_gate,
  test_alpha = .025,
  test_groups = list(1:2, 3:4),
  test_types = c("s", "p"),
  test_corr = diag(4),
  sim_n = 1e5,
  sim_success = 1
)

}
