% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_subgraphs.R
\name{calc_subgraphs}
\alias{calc_subgraphs}
\title{Generate the weighted subgraphs of a multiple comparison procedure graph}
\usage{
calc_subgraphs(g)
}
\description{
Generate the weighted subgraphs of a multiple comparison procedure graph
}
