% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPosteriorProbsBinom.R
\name{CalcPosteriorProbsBinom}
\alias{CalcPosteriorProbsBinom}
\title{CalcPosteriorProbsBinom}
\usage{
CalcPosteriorProbsBinom(
  dParam1S,
  dParam2S,
  dParam1E,
  dParam2E,
  dDelta1,
  dDelta2
)
}
\description{
{ \ifelse{html}{ This function assumes \eqn{\pi}\out{<sub>E</sub>} ~ Beta( dParam1E, dParam2E ) and
\eqn{\pi}\out{<sub>S</sub>}~ Beta( dParam1S, dParam2S ) and computes
Pr( \eqn{\pi}\out{<sub>E</sub>} - \eqn{\pi}\out{<sub>S</sub> } > dDelta1 ) and
Pr( \eqn{\pi}\out{<sub>E</sub>}-\eqn{\pi}\out{<sub>S</sub>} > dDelta2 ) }{This function assumes \eqn{\pi_E} ~ Beta( dParam1E, dParam2E )
and  \eqn{\pi_S}~ Beta( dParam1S, dParam2S ) and computes Pr( \eqn{\pi_E - \piS} > dDelta1 ) and and Pr( \eqn{\pi_E-\pi_S} > dDelta2 )}
}
}
