% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_has_missing.R
\name{corr_has_missing}
\alias{corr_has_missing}
\title{Check if a correlation matrix has any missing values in a given subset}
\usage{
corr_has_missing(corr, indices)
}
\arguments{
\item{corr}{A numeric matrix}

\item{indices}{A numeric vector indicating which rows and columns to check in
for missing values}
}
\description{
Check if a correlation matrix has any missing values in a given subset
}
\examples{
corr <- matrix(nrow = 4, ncol = 4)
corr[3:4, 3:4] <- .5
diag(corr) <- 1

corr_has_missing(corr, 2:3)
# TRUE

corr_has_missing(corr, 3:4)
# FALSE
}
