% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_graph.R
\name{create_graph}
\alias{create_graph}
\title{\code{initial_graph} object}
\usage{
create_graph(hypotheses, transitions, names = NULL)
}
\arguments{
\item{hypotheses}{A numeric vector of hypothesis weights in an initial
graphical multiple comparison procedure. Must be a vector of values between
0 & 1 (inclusive). The length should match the row and column lengths of
\code{transitions}. The sum of hypothesis weights cannot exceed 1}

\item{transitions}{A numeric matrix of transition weights between hypotheses
in an initial graphical multiple comparison procedure. Must be a square
matrix of values between 0 & 1 (inclusive). The row and column lengths
should match the length of \code{hypotheses}. Each row (Transition weights
leaving a hypothesis) can sum to no more than 1. The diagonal (Transition
weights from a hypothesis to itself) must be all 0s}

\item{names}{(Optional) A character vector of hypothesis names. If not
provided, names from \code{hypotheses} and \code{transitions} will be used. If names
are not specified, hypotheses will be named sequentially as H1, H2, ...}
}
\value{
An S3 object of class \code{initial_graph}
}
\description{
Creates a list that represents an initial graphical multiple comparison
procedure.
}
\examples{
# A graphical multiple comparison procedure with two primary hypotheses (H1
# and H2) and two secondary hypotheses (H3 and H4)
# See Figure 1 in Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer,
# W., & Rohmeyer, K. (2011). Graphical approaches for multiple comparison
# procedures using weighted Bonferroni, Simes, or parametric tests.
# Biometrical Journal, 53(6), 894-913.
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
g
}
