% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_node_fast.R
\name{delete_hypothesis}
\alias{delete_hypothesis}
\title{Remove a hypothesis from a graph}
\usage{
delete_hypothesis(graph, del_index)
}
\arguments{
\item{graph}{An object of class \code{initial_graph}, for example as returned by
\code{create_graph()}}

\item{del_index}{An integer scalar indicating which hypothesis to remove}
}
\value{
An object of class \code{initial_graph}
}
\description{
Remove a hypothesis from a graph
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- create_graph(hypotheses, transitions)

delete_hypothesis(g, 3)
}
