% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_pairs.R
\name{edge_pairs}
\alias{edge_pairs}
\title{Find pairs of vertices that are connected in both directions}
\usage{
edge_pairs(graph)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=graph_create]{graph_create()}}.}
}
\value{
A list of vertex pairs which are connected in both directions. NULL
if no such pairs are found.
}
\description{
For an initial graph, find pairs of hypotheses that are connected in both
directions. This is used to plot graphs using \code{\link[=plot.initial_graph]{plot.initial_graph()}}.
}
\keyword{internal}
