% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graphs.R
\name{bonferroni}
\alias{bonferroni}
\alias{bonferroni_holm}
\alias{huque_alosh_bhore_2011}
\alias{wiens_dmitrienko_2005}
\alias{fixed_sequence}
\alias{fallback}
\alias{simple_successive_1}
\alias{simple_successive_2}
\alias{random_graph}
\title{Example graphs from a selection of papers}
\usage{
bonferroni(num_hyps, hyp_names = NULL)

bonferroni_holm(num_hyps, hyp_names = NULL)

huque_alosh_bhore_2011(hyp_names = NULL)

wiens_dmitrienko_2005(hyp_names = NULL)

fixed_sequence(num_hyps = 3, hyp_names = NULL)

fallback(hypotheses = c(1, 0, 0), hyp_names = NULL)

simple_successive_1(hyp_names = NULL)

simple_successive_2(hyp_names = NULL)

random_graph(num_hyps, hyp_names = NULL)
}
\arguments{
\item{num_hyps}{Number of vertices in the graph}

\item{hyp_names}{(Optional) A character vector of hypothesis names. If names
are not specified, hypotheses will be named sequentially as H1, H2, ...}

\item{hypotheses}{Hypothesis weights in a fallback procedure}
}
\value{
An S3 object as returned by \code{\link[=graph_create]{graph_create()}}
}
\description{
Example graphs from a selection of papers
}
\examples{
bonferroni_holm(3, hyp_names = paste("dose", letters[1:3]))
}
