% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graphs.R
\name{bonferroni}
\alias{bonferroni}
\alias{bonferroni_holm}
\alias{huque_etal}
\alias{fallback}
\alias{fallback_improved_1}
\alias{fallback_improved_2}
\alias{fixed_sequence}
\alias{simple_successive_1}
\alias{simple_successive_2}
\alias{random_graph}
\alias{two_doses_two_primary_two_secondary}
\alias{three_doses_two_primary_two_secondary}
\title{Example graphs of commonly used multiple comparison procedures}
\usage{
bonferroni(hypotheses, hyp_names = NULL)

bonferroni_holm(hypotheses, hyp_names = NULL)

huque_etal(hyp_names = NULL)

fallback(hypotheses, hyp_names = NULL)

fallback_improved_1(hypotheses, hyp_names = NULL)

fallback_improved_2(hypotheses, epsilon = 1e-04, hyp_names = NULL)

fixed_sequence(num_hyps, hyp_names = NULL)

simple_successive_1(hyp_names = NULL)

simple_successive_2(hyp_names = NULL)

random_graph(num_hyps, hyp_names = NULL)

two_doses_two_primary_two_secondary(hyp_names = NULL)

three_doses_two_primary_two_secondary(hyp_names = NULL)
}
\arguments{
\item{hypotheses}{(Optional) A numeric vector of hypothesis weights in a
graphical multiple comparison procedure. Must be a vector of values
between 0 & 1 (inclusive). The length should match \code{num_hyps} and the
length of \code{hyp_names}. The sum of hypothesis weights should not exceed 1.}

\item{hyp_names}{(Optional) A character vector of hypothesis names. The
length should match \code{num_hyps} and the length of \code{hypotheses}. If
\code{hyp_names} are not specified, hypotheses will be named sequentially as
H1, H2, .......}

\item{epsilon}{(Optional) A numeric scalar indicating the value of the
\eqn{\epsilon} edge. This should be a much smaller value than hypothesis
and transition weights. The default is 1e-4.}

\item{num_hyps}{(Optional) Number of hypotheses in a graphical multiple
comparison procedure.}
}
\value{
An S3 object as returned by \code{\link[=graph_create]{graph_create()}}.
}
\description{
Built-in functions to quickly generate select graphical multiple comparison
procedures.
}
\examples{
# Bretz et al. (2009)
bonferroni(hypotheses = rep(1 / 3, 3))
# Bretz et al. (2009)
bonferroni_holm(hypotheses = rep(1 / 3, 3))
# Huque et al. (2011)
huque_etal()
# Wiens (2003)
fallback(rep(1 / 3, 3))
# Wiens and Dmitrienko (2005)
fallback_improved_1(rep(1 / 3, 3))
# Bretz et al. (2009)
fallback_improved_2(rep(1 / 3, 3))
# Maurer et al. (1995); Westfall and Krishen (2001)
fixed_sequence(3)
# Figure 1 in Bretz et al. (2011)
simple_successive_1()
# Figure 4 in Bretz et al. (2011)
simple_successive_2()
# Create a random graph with three hypotheses
random_graph(3)
# Figure 6 in Xi and Bretz et al. (2019)
two_doses_two_primary_two_secondary()
# Add another dose to Figure 6 in Xi and Bretz et al. (2019)
three_doses_two_primary_two_secondary()
}
\references{
\itemize{
\item \insertRef{bretz-2009-graphical}{graphicalMCP}
\item \insertRef{bretz-2011-graphical}{graphicalMCP}
\item \insertRef{huque-2011-addressing}{graphicalMCP}
\item \insertRef{maurer-1995-multiple}{graphicalMCP}
\item \insertRef{westfall-2001-optimally}{graphicalMCP}
\item \insertRef{wiens-2003-fixed}{graphicalMCP}
\item \insertRef{wiens-2005-fallback}{graphicalMCP}
\item \insertRef{xi-2019-symmetric}{graphicalMCP}
}
}
\seealso{
\code{\link[=graph_create]{graph_create()}} for a general way to create the initial graph.

Other initial graphs: 
\code{\link{as_initial_graph}()},
\code{\link{graph_create}()},
\code{\link{plot.initial_graph}()},
\code{\link{print.initial_graph}()}
}
\concept{initial graphs}
