% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_graph.R
\name{gamma_graph}
\alias{gamma_graph}
\title{Create a graph with gamma edges}
\usage{
gamma_graph(graph, gamma_props)
}
\arguments{
\item{gamma_props}{}
}
\value{
A gamma graph function, which accepts a vector of gammas, with one
gamma value per hypothesis. The gamma value for each hypothesis is spread
across transitions according to \code{gamma_props}
}
\description{
Create a graph with gamma edges
}
\examples{
successive1 <- simple_successive_1()
successive2 <- simple_successive_2()

gamma_props <- rbind(
  c(0, 1, -1, 0),
  c(1, 0, 0, -1),
  c(0, 0, 0, 0),
  c(0, 0, 0, 0)
)

gamma1 <- c(.999, .999, 0, 0)
gamma2 <- c(.5, .5, 0, 0)

successive_gamma <- gamma_graph(successive1, gamma_props)

successive_gamma(gamma1)

all.equal(successive2, successive_gamma(gamma2))

}
