% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_generate.R
\name{generate_weights}
\alias{generate_weights}
\title{Generate weights for the full closure tree of an MCP graph}
\usage{
generate_weights(graph)
}
\arguments{
\item{graph}{An MCP graph as created by \code{graph()}}
}
\value{
Returns a matrix similar to that of \code{gMCP::generateWeights}: The
first half of columns contains indices for which nodes are left. The second
half of columns contains the weights of the full closure tree of a graph
describing a multiple comparison procedure
}
\description{
Generate weights for the full closure tree of an MCP graph
}
\examples{

ex_graph <- create_graph(
  hypotheses = c(.5, .5, 0, 0),
  transitions = rbind(
    c(0,0,1,0),
    c(0,0,0,1),
    c(0,1,0,0),
    c(1,0,0,0)
  )
)

generate_weights(ex_graph)

}
