% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_hyp_weights.R
\name{get_hyp_weights}
\alias{get_hyp_weights}
\title{Get hypothesis weights for a multiple comparison procedure graph}
\usage{
get_hyp_weights(g)
}
\description{
Get hypothesis weights for a multiple comparison procedure graph
}
