% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_subgraphs.R
\name{get_subgraphs}
\alias{get_subgraphs}
\title{Get the weighted subgraphs of a multiple comparison procedure graph}
\usage{
get_subgraphs(g)
}
\description{
Get the weighted subgraphs of a multiple comparison procedure graph
}
