% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_trns_weights.R
\name{get_trns_weights}
\alias{get_trns_weights}
\title{Get transition weights for a multiple comparison procedure graph}
\usage{
get_trns_weights(g)
}
\description{
Get transition weights for a multiple comparison procedure graph
}
