% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getJavaInfo}
\alias{getJavaInfo}
\title{Get Memory and Runtime Info from JVM}
\usage{
getJavaInfo(memory = TRUE, filesystem = TRUE, runtime = TRUE)
}
\arguments{
\item{memory}{Logical whether to include memory information + number of available cores}

\item{filesystem}{Logical whether to include filesystem information (Total, free and usable space)}

\item{runtime}{Logical whether to include runtime information (Class Path, Library Path, Input Arguments)}
}
\value{
character vector of length 1 containing the memory and runtime info.
}
\description{
Get Memory and Runtime Info from JVM
}
\examples{
\dontrun{
cat(getJavaInfo())
}

}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
