% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph.R
\name{graph}
\alias{graph}
\title{\code{mcp_graph} object}
\usage{
graph(hypotheses, transitions, names = NULL)
}
\arguments{
\item{hypotheses}{A numeric vector of hypothesis weights in the multiple
comparison procedure graph. Must be a vector of values between 0 & 1
(inclusive) with length matching the row and column lengths of
\code{transitions}. The sum of hypothesis weights cannot exceed 1}

\item{transitions}{A numeric matrix of transition weights between hypotheses
in the multiple comparison procedure graph. Must be a square matrix of
values between 0 & 1 (inclusive) with row and column lengths matching the
length of \code{hypotheses}. Each row (Transition weights leaving a hypothesis)
can sum to no more than 1, and the diagonal (Transition weights from a
hypothesis to itself) must be all 0s}

\item{names}{(Optional) A character vector of hypothesis names. If not
provided, names from the weights arguments will be used. If the weights
arguments are unnamed, hypotheses will be named sequentially}

\item{graph}{An object to test}
}
\value{
An S3 object of class \code{mcp_graph}
}
\description{
Creates a list that represents a multiple comparison procedure graph
}
\examples{
graph(c(.5, .5), rbind(c(0, 1), c(1, 0)), c("doseA", "doseB"))

# Explicit names override names in the weights arguments (with a warning)
graph(c(d1 = .5, d2 = .5), rbind(c(0, 1), c(1, 0)), c("doseA", "doseB"))

# When names are not specified, hypotheses are numbered
graph(
  c(.333333, .333333, .333333),
  matrix(c(0, .5, .5, .5, 0, .5, .5, .5, 0), nrow = 3, byrow = TRUE),
)
}
