% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generate_weights.R
\name{graph_generate_weights}
\alias{graph_generate_weights}
\title{Calculate weights for the closure of a graph}
\usage{
graph_generate_weights(graph)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=graph_create]{graph_create()}}}
}
\value{
A numeric matrix of all intersection hypothesis weights. Each row
corresponds to a single intersection hypothesis. The first half of the
columns indicate which hypotheses are included in the given intersection
hypothesis, and the second half of columns are the weights.
}
\description{
The closure of a graph is the set of all sub-graphs, or intersections
hypotheses, of a graph. Weights for each sub-graph are calculated using the
weighting strategy defined in Bretz et al (2011).
}
\section{Performance}{

Much thought was given to the performance of this code, as the memory and
time usage can grow quickly as graph size grows. On the systems used for
testing, a size 10 graph had a median run time of 20-60 ms. Run time
increases at a rate of O(2 ^ n), so e.g. a size 5 graph takes approximately
twice as long to run as a size 4 graph. See \code{vignette("generate-closure")}
for more information about using the closure, including performance metrics.
}

\examples{

par_gate <- graph_create(
  hypotheses = c(.5, .5, 0, 0),
  transitions = rbind(
    c(0, 0, 1, 0),
    c(0, 0, 0, 1),
    c(0, 1, 0, 0),
    c(1, 0, 0, 0)
  )
)

graph_generate_weights(par_gate)

}
