% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_update.R
\name{graph_update}
\alias{graph_update}
\title{Delete hypotheses from a graph}
\usage{
graph_update(graph, delete)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=graph_create]{graph_create()}}}

\item{delete}{A logical or integer vector, denoting which hypotheses to
keep/delete. A logical vector must match the size of the graph, with one
entry per hypothesis, and it results in "unordered mode," where selected
hypotheses are deleted in sequential order. An integer vector can be any
length, but must only contain valid hypothesis numbers (greater than 0, and
less than or equal to the size of the graph). This will trigger "ordered
mode," where selected hypotheses are deleted in the order they appear in
\code{delete}}
}
\value{
An object of class \code{updated_graph} with 4 elements
\itemize{
\item The initial graph object
\item The updated graph object with specified hypotheses deleted
\item A numeric vector indicating which hypotheses were deleted
\item When using ordered mode, a list of intermediate graphs for each deletion
step
}
}
\description{
It is not always obvious what a graph's weights will look like after deleting
one or more hypotheses. While \code{\link[=graph_generate_weights]{graph_generate_weights()}} calculates all
sub-graphs' hypothesis weights, \code{graph_update()} gives a more detailed view
for a single set of deletions, including transition weights.
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- graph_create(hypotheses, transitions)

# Delete the second and third hypotheses in any order
graph_update(g, c(FALSE, TRUE, TRUE, FALSE))

# Equivalent in ordered mode
graph_update(g, 2:3)
}
