% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalMCP-package.R
\docType{package}
\name{graphicalMCP-package}
\alias{graphicalMCP}
\alias{graphicalMCP-package}
\title{graphicalMCP: Graphical Multiple Comparison Procedures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Multiple comparison procedures (MCPs) control the familywise error rate in clinical trials. Graphical MCPs include many commonly used procedures as special cases; see Bretz et al. (2011) \doi{10.1002/bimj.201000239}, Lu (2016) \doi{10.1002/sim.6985}, and Xi et al. (2017) \doi{10.1002/bimj.201600233}. This package is a low-dependency implementation of graphical MCPs which allow mixed types of tests. It also includes power simulations and visualization of graphical MCPs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openpharma.github.io/graphicalMCP/}
  \item \url{https://github.com/openpharma/graphicalMCP}
  \item Report bugs at \url{https://github.com/openpharma/graphicalMCP/issues}
}

}
\author{
\strong{Maintainer}: Dong Xi \email{dong.xi1@gilead.com}

Authors:
\itemize{
  \item Ethan Brockmann \email{ethan.brockmann@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences, Inc. [copyright holder, funder]
}

}
\keyword{internal}
