% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph.R
\name{is_optimal}
\alias{is_optimal}
\title{Check if a graph is optimal}
\usage{
is_optimal(graph)
}
\arguments{
\item{graph}{An \code{mcp_graph} object as created by \code{graph()}}
}
\value{
A Boolean result specifying whether the graph is optimal or not
}
\description{
Check if a graph is optimal
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0))
g <- graph(hypotheses, transitions)

is_optimal(g)
# TRUE

g$hypotheses[[1]] <- 0.4
is_optimal(g)
# FALSE
}
