% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph.R
\name{mcp_graph}
\alias{mcp_graph}
\alias{is_mcp_graph}
\title{\code{mcp_graph} vector}
\usage{
mcp_graph(hyp_names = NULL, n_hyps = NULL)

is_mcp_graph(g)
}
\arguments{
\item{hyp_names}{A vector of hypothesis names for nodes in the multiple
comparison procedure graph. Defaults to hypothesis numbers if no names are
provided}

\item{n_hyps}{(Optional) An integer specifying the number of nodes in the
multiple comparison procedure graph}

\item{g}{An object to test}
}
\value{
An S3 vector of class \code{mcp_graph}.
}
\description{
This creates an integer vector that represents the hypotheses of a multiple
comparison procedure graph.
}
\examples{
mcp_graph(c("doseA", "doseB"))

mcp_graph(n_hyps = 5)
}
