% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_report_print2.R
\name{print.graph_report2}
\alias{print.graph_report2}
\title{S3 print method for class \code{graph_report2}}
\usage{
\method{print}{graph_report2}(x, ...)
}
\arguments{
\item{x}{An object of class \code{graph_report} to print}

\item{...}{Other values passed on to other methods (currently unused)}
}
\description{
A graph report (2) was originally written for the separate critical value vs
adjusted-p method functions, and it displays
\itemize{
\item The initial graph being tested,
\item p-values & alpha used for tests,
\item Which hypotheses can be rejected, and
\item Detailed test results matrix, including the results of
\code{generate_weights()} & test results for each intersection hypothesis
}
}
