% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_print.R
\name{print.initial_graph}
\alias{print.initial_graph}
\title{S3 print method for the class 'initial_graph'}
\usage{
\method{print}{initial_graph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{initial_graph} to print}

\item{...}{Other values passed on to other methods (currently unused)}
}
\description{
A printed \code{initial_graph} displays a header stating what the object is, the
hypothesis weights, and the transition weights.
}
