% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_print.R
\name{print.mcp_graph}
\alias{print.mcp_graph}
\title{S3 print method for class 'mcp_graph'}
\usage{
\method{print}{mcp_graph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcp_graph} to print}

\item{...}{Other values passed on to other methods (currently unused)}
}
\description{
A printed \code{mcp_graph} displays a header stating what the object is, the
hypothesis weights, and the transition weights.
}
