% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.updated_graph.R
\name{print.updated_graph}
\alias{print.updated_graph}
\title{S3 print method for the class \code{updated_graph}}
\usage{
\method{print}{updated_graph}(x, ..., precision = 6, indent = 2)
}
\arguments{
\item{x}{An object of class \code{updated_graph} to print}

\item{...}{Other values passed on to other methods (currently unused)}

\item{precision}{An integer scalar indicating the number of significant
figures to include in numeric values}

\item{indent}{An integer scalar indicating how many spaces to indent results}
}
\description{
A printed \code{updated_graph} displays the initial graph, the updated graph, and
the sequence of graphs between the two (if it's available).
}
