% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_hyp_weights.R
\name{set_hyp_weights}
\alias{set_hyp_weights}
\title{Set hypothesis weights for a multiple comparison procedure graph}
\usage{
set_hyp_weights(g, weights)
}
\description{
Set hypothesis weights for a multiple comparison procedure graph
}
