% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_graph_trns_weights.R
\name{set_trns_weights}
\alias{set_trns_weights}
\title{Set transition weights for a multiple comparison procedure graph}
\usage{
set_trns_weights(g, weights)
}
\description{
Set transition weights for a multiple comparison procedure graph
}
