% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyGraphs.R
\name{subgraph}
\alias{subgraph}
\title{Get a subgraph}
\usage{
subgraph(graph, subset)
}
\arguments{
\item{graph}{A graph of class \code{\link{graphMCP}}.}

\item{subset}{A logical or character vector specifying the nodes in the
subgraph.}
}
\value{
A subgraph containing only the specified nodes.
}
\description{
Given a set of nodes and a graph this function creates the subgraph
containing only the specified nodes.
}
\examples{


graph <- improvedParallelGatekeeping()
subgraph(graph, c(TRUE, FALSE, TRUE, FALSE))
subgraph(graph, c("H1", "H3"))


}
\seealso{
\code{\link{graphMCP}}
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{graphs}
\keyword{print}
