% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_bonferroni}
\alias{test_bonferroni}
\alias{test_parametric}
\alias{test_simes}
\title{Test hypotheses with the \verb{p <= (c * ) w * a} method}
\usage{
test_bonferroni(p_values, weights, alpha, verbose = TRUE)

test_parametric(p_values, weights, alpha, corr, verbose = TRUE)

test_simes(p_values, weights, alpha, verbose = TRUE)
}
\arguments{
\item{p_values}{A numeric vector of p-values}

\item{weights}{A numeric vector of hypothesis weights}

\item{alpha}{A numeric scalar specifying the level to test weighted p-values
against}

\item{verbose}{A logical scalar specifying how detailed the results should be}
}
\value{
A vector of test results if \code{verbose} is FALSE, otherwise a data
frame with detailed test values and results
}
\description{
Test hypotheses with the \verb{p <= (c * ) w * a} method
}
