% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_graph.R
\name{test_graph}
\alias{test_graph}
\title{Apply weighted Bonferroni, parametric, and Simes tests}
\usage{
test_graph(
  graph,
  p_values,
  alpha = 0.05,
  corr = NULL,
  use_cJ = FALSE,
  tests = list(bonferroni = list(seq_along(graph$hypotheses)), parametric = NULL, simes =
    NULL)
)
}
\arguments{
\item{graph}{An initial graph as created by \code{create_graph()}}

\item{p_values}{A numeric vector of p-values, one for each hypothesis in the
graph}

\item{alpha}{A numeric vector of length 1 specifying the un-weighted alpha
level at which to test each hypothesis}

\item{tests}{A list with three elements, \code{bonferroni}, \code{simes}, and
\code{parametric}. Each element is a list of hypothesis groups to apply the
given test to. Each hypothesis must be specified exactly once, so that the
length of all elements of the list equals the number of hypotheses. The
default is to apply the weighted Bonferroni test to all hypotheses}
}
\value{
A \code{graph_report} object, consisting of
\itemize{
\item The initial graph being tested,
\item p-values & alpha used for tests,
\item Which hypotheses can be rejected, and
\item Detailed test results matrix, including the results of
\code{generate_weights()} & test results for each intersection hypothesis
}
}
\description{
Apply weighted Bonferroni, parametric, and Simes tests
}
\examples{

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)

test_graph(
  g,
  p_values = c(.01, .02, .05, .1),
  alpha = .05,
  tests = list(
    bonferroni = 1,
    simes = list(c(2)),
    parametric = list(c(3, 4))
  )
)
}
