% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_report_adjusted_p.R
\name{test_graph2}
\alias{test_graph2}
\title{Title}
\usage{
test_graph2(
  graph,
  p_values,
  alpha = 0.05,
  groups = list(seq_along(graph$hypotheses)),
  tests = c("bonferroni"),
  corr = NULL,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
\examples{

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- create_graph(hypotheses, transitions)
p <- c(.01, .005, .015, .022)

corr <- matrix(nrow = 4, ncol = 4)
corr[3:4, 3:4] <- .5
diag(corr) <- 1

corr2 <- matrix(.5, nrow = 4, ncol = 4)
diag(corr2) <- 1

# The default is all Bonferroni with alpha = .05
test_graph2(g, p)

# But tests can be specified at the hypothesis-level
test_graph2(
  graph = g,
  p_values = p,
  alpha = .025,
  groups = list(1, 2, 3:4),
  tests = c("bonferroni", "simes", "parametric"),
  corr = corr
)
}
