% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_test_closure.R, R/graph_test_shortcut.R
\name{graph_test_closure}
\alias{graph_test_closure}
\alias{graph_test_shortcut}
\title{Report details of hypothesis rejections}
\usage{
graph_test_closure(
  graph,
  p,
  alpha = 0.025,
  test_groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  test_corr = rep(list(NA), length(test_types)),
  adjust_p_custom = NULL,
  addl_args_custom = list(),
  verbose = FALSE,
  test_values = FALSE
)

graph_test_shortcut(
  graph,
  p,
  alpha = 0.025,
  verbose = FALSE,
  test_values = FALSE
)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=graph_create]{graph_create()}}}

\item{p}{A numeric vector of p-values}

\item{alpha}{A numeric scalar specifying the significance level for testing}

\item{test_groups}{A list of numeric vectors specifying hypotheses to test
together. Part 1 of the test specification.}

\item{test_types}{A character vector of tests to apply to the test groups.
Part 2 of the test specification.}

\item{test_corr}{(Optional) A list of numeric matrices. Part 3 of the test
specification. Each test group must have exactly one entry in the list:
Bonferroni and Simes test groups should have entries of NULL, and
parametric test groups should have numeric matrices specifying the (known
or estimated) pairwise correlations between the test statistics of all
hypotheses in the group.}

\item{adjust_p_custom}{(Optional) A function that implements a new testing
algorithm. The inputs and outputs must match the input and output structure
of the other \verb{adjust_p_*} functions}

\item{addl_args_custom}{(Optional) A list of additional arguments passed
on when \code{adjust_p_custom} is called}

\item{verbose}{A logical scalar specifying whether the details of the
adjusted p-value calculations should be included in results}

\item{test_values}{A logical scalar specifying whether details of the
adjusted significance calculations should be included in results}
}
\value{
A \code{graph_report} object, a list of 4 elements: \code{inputs}, \code{outputs},
\code{details}, and \code{test_values}
\itemize{
\item Inputs - A list of the input parameters used to run the test
\item Outputs - A list of global test results
\item Details - A list of detailed adjusted p-value calculations (graph
deletion sequence for shortcut testing)
\item Test values - A list with hypothesis-level test details for each
intersection using the adjusted significance method (Details about each
step taken for shortcut testing)
}
}
\description{
The graph testing functions apply a specified test strategy to a graph and a
set of p-values, giving rich information on which hypotheses are significant,
and why. Results include hypothesis rejection decisions, but also the test
values that led to the final result. The functions include options for
reporting details using the adjusted p-value method or adjusted significance
method.
}
\details{
The test specification (\code{test_groups}, \code{test_types}, and \code{test_corr}) can be
specified either named or unnamed. If unnamed, it's assumed that all 3 are
ordered the same way, i.e. the nth elements of \code{test_types} and \code{test_corr}
apply to the nth group in \code{test_groups}. Naming each element with consistent
names across the three vectors will mean that they do not have to retain the
same order, and also may be a more robust way to keep track of testing
details.
}
\examples{

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- graph_create(hypotheses, transitions)
p <- c(.01, .005, .015, .022)

corr_list <- list(NA, matrix(c(1, .5, .5, 1), nrow = 2, byrow = TRUE))

# The default is all Bonferroni with alpha = .025
graph_test_closure(g, p)

# But tests can be specified at the hypothesis-level
graph_test_closure(
  graph = g,
  p = p,
  alpha = .025,
  test_groups = list(1:2, 3:4),
  test_types = c("bonferroni", "parametric"),
  test_corr = corr_list
)
}
