% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_graph_plain.R
\name{update_graph_plain}
\alias{update_graph_plain}
\title{Delete multiple hypotheses from a graph}
\usage{
update_graph_plain(graph, keep)
}
\arguments{
\item{graph}{An MCP graph as created by \code{initial_graph()}}

\item{keep}{A Boolean vector denoting which hypotheses to delete. An entry of
\code{FALSE} corresponds to a deletion, and \code{TRUE} corresponds to keeping a
hypothesis}
}
\value{
An object of class \code{updated_graph} with 3 elements
\itemize{
\item The initial graph object
\item The Boolean vector of keep/delete indicators
\item The updated graph object with appropriate hypotheses removed
}
}
\description{
\code{update_graph_plain()} seeks to stay as simple as possible, and as similar as
possible to the algorithm laid out in the paper. Hypotheses and
transitions are directly updated using the formulas from the paper and
vectorized operations.
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(c(0, 0, 1, 0),
                     c(0, 0, 0, 1),
                     c(0, 1, 0, 0),
                     c(1, 0, 0, 0))
g <- create_graph(hypotheses, transitions)

# Delete the third hypothesis
update_graph(g, c(TRUE, TRUE, FALSE, TRUE))

#

}
