% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgt_generate.R
\name{wgt_generate}
\alias{wgt_generate}
\title{Generate weights for the full closure tree of a graph}
\usage{
wgt_generate(g, w)
}
\arguments{
\item{g}{A graph defined as a matrix where entry \verb{i, j} corresponds to the
transition weight between hypotheses \code{i} and \code{j}}

\item{w}{A vector of weights corresponding to the hypotheses in the graph}
}
\value{
Returns a matrix similar to that of \code{gMCP::generateWeights}: The
first half of columns contains indices for which nodes are left. The second
half of columns contains the weights of the full closure tree of a graph
describing a multiple comparison procedure
}
\description{
Generate weights for the full closure tree of a graph
}
\examples{

# Example taken from gMCP::generateWeights documentation
g <- matrix(
  c(
    0,
    0,
    1,
    0,
    0,
    0,
    0,
    1,
    0,
    1,
    0,
    0,
    1,
    0,
    0,
    0
  ),
  nrow = 4,
  byrow = TRUE
)

## Choose weights
w <- c(.5, .5, 0, 0)

generate_weights(g, w)

}
