% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_p.R
\name{adjust_p_bonferroni}
\alias{adjust_p_bonferroni}
\alias{adjust_p_parametric}
\alias{adjust_p_simes}
\title{Calculate adjusted p-values}
\usage{
adjust_p_bonferroni(p, hypotheses, group)

adjust_p_parametric(p, hypotheses, group, test_corr = NULL)

adjust_p_simes(p, hypotheses, group)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{hypotheses}{A numeric vector of hypothesis weights}

\item{group}{A numeric vector used to subset p-values, hypotheses, and any
other arguments}

\item{test_corr}{(Optional) A numeric matrix of correlations between
hypotheses' test statistics}
}
\value{
A single adjusted p-value for the given group
}
\description{
Currently only Bonferroni, parametric, and Simes adjusted p-value
calculations are available in this package. However, there is a mechanism for
users to insert their own custom test type by defining a new adjusted p-value
function, which matches the format of existing ones. See
\code{vignette("custom-tests")} for more details.
}
\examples{
set.seed(22723)

w <- c("H1" = .75, "H2" = .25, "H3" = 0)
p <- c("H1" = .019, "H2" = .025, "H3" = .05)

adjust_p_bonferroni(p, w)
adjust_p_simes(p, w)

corr1 <- diag(3)
corr2 <- corr1
corr2[lower.tri(corr2)] <- corr2[upper.tri(corr2)] <- runif(3, -1, 1)

# No correlation
adjust_p_parametric(p, w, corr1)

# Uniform random pairwise correlations
adjust_p_parametric(p, w, corr2)
}
\keyword{internal}
