% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_graph.R
\name{analyze_graph}
\alias{analyze_graph}
\title{Check whether a graph is valid and optimal}
\usage{
analyze_graph(graph)
}
\arguments{
\item{graph}{An \code{mcp_graph} object as created by \code{graph()}}
}
\value{
A character vector, either "optimal", "valid", or "invalid"
}
\description{
Check whether a graph is valid and optimal
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(c(0, 0, 1, 0),
                     c(0, 0, 0, 1),
                     c(0, 1, 0, 0),
                     c(1, 0, 0, 0))
g <- graph(hypotheses, transitions)

analyze_graph(g)
# "optimal"

# Hypothesis weights no longer sum to 1
g$hypotheses[[1]] <- 0.4
analyze_graph(g)
# "valid"

# Names of hypotheses and transitions no longer match
names(g$hypotheses) <- 1:4
analyze_graph(g)
# "invalid"
}
