% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_graph.R
\name{as_initial_graph}
\alias{as_initial_graph}
\alias{as_initial_graph.graphMCP}
\alias{as_initial_graph.igraph}
\alias{as_graphMCP}
\alias{as_graphMCP.initial_graph}
\alias{as_igraph}
\alias{as_igraph.initial_graph}
\title{Convert between gMCP and graphicalMCP graph classes}
\usage{
as_initial_graph(graph)

\method{as_initial_graph}{graphMCP}(graph)

\method{as_initial_graph}{igraph}(graph)

as_graphMCP(graph)

\method{as_graphMCP}{initial_graph}(graph)

as_igraph(graph)

\method{as_igraph}{initial_graph}(graph)
}
\arguments{
\item{graph}{An \code{initial_graph} object from graphicalMCP, a \code{graphMCP} object
from gMCP, or an \code{igraph} object from igraph, depending on which direction
you're converting}
}
\value{
For \code{as_graphMCP()}, a gMCP graph object, for \code{as_igraph()}, an
igraph object, and for \code{as_initial_graph()}, a graphicalMCP graph object
}
\description{
Convert between gMCP and graphicalMCP graph classes
}
\examples{
g1 <- random_graph(5)

if (requireNamespace("gMCP", quietly = TRUE)) {
  g2 <- as_graphMCP(g1)

  all.equal(g1, as_initial_graph(g2))
}

if (requireNamespace("igraph", quietly = TRUE)) {
  g3 <- as_igraph(g1)

  all.equal(g1, as_initial_graph(g3))
}
}
