% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_tests.R
\name{bonferroni_test_vals}
\alias{bonferroni_test_vals}
\alias{parametric_test_vals}
\alias{simes_test_vals}
\title{Test hypotheses with the critical values method}
\usage{
bonferroni_test_vals(p, weights, alpha)

parametric_test_vals(p, weights, alpha, corr)

simes_test_vals(p, weights, alpha)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{weights}{A numeric vector of hypothesis weights}

\item{alpha}{A numeric scalar specifying the global significance level for
testing}

\item{corr}{A numeric matrix of correlations between hypotheses' test
statistics}
}
\value{
A data frame with columns specifying the values used to calculate
each hypothesis test
}
\description{
Test hypotheses with the critical values method
}
\examples{
w <- c(H1 = .5, H2 = .5, H3 = 0, H4 = 0)

p <- c(.024, .01, .026, .027)

graphicalMCP:::bonferroni_test_vals(p, w, .05)
graphicalMCP:::parametric_test_vals(p, w, .05, diag(4))
graphicalMCP:::simes_test_vals(p, w, .05)
}
\keyword{(2009).}
\keyword{(2011).}
\keyword{(2016).}
\keyword{(2017).}
\keyword{(2019).}
\keyword{1489–1501.}
\keyword{28(4),}
\keyword{30(13),}
\keyword{35(22),}
\keyword{38(27),}
\keyword{4041–4055.}
\keyword{5268–5282.}
\keyword{53(6),}
\keyword{586–604.}
\keyword{59(5),}
\keyword{894–913.}
\keyword{918–931.}
\keyword{A}
\keyword{Biometrical}
\keyword{Bonferroni}
\keyword{Bonferroni,}
\keyword{Brannath,}
\keyword{Bretz,}
\keyword{D.,}
\keyword{E.,}
\keyword{F.}
\keyword{F.,}
\keyword{G.}
\keyword{Glimm,}
\keyword{Graphical}
\keyword{Hochberg}
\keyword{Hommel,}
\keyword{Journal,}
\keyword{K.}
\keyword{Klinglmueller,}
\keyword{Lu,}
\keyword{M.}
\keyword{M.,}
\keyword{Maurer,}
\keyword{Medicine,}
\keyword{Posch,}
\keyword{Rohmeyer,}
\keyword{Simes}
\keyword{Simes,}
\keyword{Statistics}
\keyword{Symmetric}
\keyword{Test}
\keyword{W.,}
\keyword{Xi,}
\keyword{a}
\keyword{analyses}
\keyword{and}
\keyword{application}
\keyword{approach}
\keyword{approaches}
\keyword{comparison}
\keyword{considerations}
\keyword{endpoint}
\keyword{equally}
\keyword{for}
\keyword{framework}
\keyword{graphical}
\keyword{graphs}
\keyword{https://doi.org/10.1002/bimj.201000239}
\keyword{https://doi.org/10.1002/bimj.201600233}
\keyword{https://doi.org/10.1002/sim.3495}
\keyword{https://doi.org/10.1002/sim.3988}
\keyword{https://doi.org/10.1002/sim.6985}
\keyword{https://doi.org/10.1002/sim.8375}
\keyword{in}
\keyword{internal}
\keyword{mixture}
\keyword{multiple}
\keyword{of}
\keyword{or}
\keyword{parametric}
\keyword{power}
\keyword{procedure.}
\keyword{procedures}
\keyword{procedures.}
\keyword{rejective}
\keyword{sequentially}
\keyword{test}
\keyword{tests,}
\keyword{tests.}
\keyword{the}
\keyword{to}
\keyword{unified}
\keyword{using}
\keyword{weighted}
\keyword{with}
