% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_tests.R
\name{bonferroni_test_vals}
\alias{bonferroni_test_vals}
\alias{parametric_test_vals}
\alias{simes_test_vals}
\title{Test hypotheses with the \verb{p <= (c * ) w * a} method}
\usage{
bonferroni_test_vals(p_values, weights, alpha, corr = NULL)

parametric_test_vals(p_values, weights, alpha, corr = NULL)

simes_test_vals(p_values, weights, alpha, corr = NULL)
}
\arguments{
\item{p_values}{A numeric vector of p-values}

\item{weights}{A numeric vector of hypothesis weights}

\item{alpha}{A numeric scalar specifying the level to test weighted p-values
against}

\item{corr}{(Optional) A numeric matrix indicating the correlation between
the test statistics which generated the p-values. For parametric testing,
\code{corr} must be a square matrix with side length equal to the length of \code{p}
and \code{weights}. Ignored for Simes and Bonferroni testing}
}
\value{
A data frame with columns specifying the values used to calculate
each hypothesis test
}
\description{
Test hypotheses with the \verb{p <= (c * ) w * a} method
}
