% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_tests.R
\name{bonferroni_test_vals}
\alias{bonferroni_test_vals}
\alias{parametric_test_vals}
\alias{simes_test_vals}
\title{Test hypotheses with the \verb{p <= (c * ) w * a} method}
\usage{
bonferroni_test_vals(p_values, weights, alpha, corr = NULL)

parametric_test_vals(p_values, weights, alpha, corr = NULL)

simes_test_vals(p_values, weights, alpha, corr = NULL)
}
\arguments{
\item{p_values}{A numeric vector of p-values}

\item{weights}{A numeric vector of hypothesis weights}

\item{alpha}{A numeric scalar specifying the level to test weighted p-values
against}

\item{verbose}{A logical scalar specifying how detailed the results should be}
}
\value{
A vector of test results if \code{verbose} is FALSE, otherwise a data
frame with detailed test values and results
}
\description{
Test hypotheses with the \verb{p <= (c * ) w * a} method
}
