% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_graph.R
\name{create_graph}
\alias{create_graph}
\title{Define a graph representing a multiple comparison procedure}
\usage{
create_graph(hypotheses, transitions, names = NULL)
}
\arguments{
\item{hypotheses}{A numeric vector of hypothesis weights in an initial
graphical multiple comparison procedure. Must be a vector of values between
0 & 1 (inclusive). The length should match the row and column lengths of
\code{transitions}. The sum of hypothesis weights cannot exceed 1}

\item{transitions}{A numeric matrix of transition weights between hypotheses
in an initial graphical multiple comparison procedure. Must be a square
matrix of values between 0 & 1 (inclusive). The row and column lengths
should match the length of \code{hypotheses}. Each row (Transition weights
leaving a hypothesis) can sum to no more than 1. The diagonal (Transition
weights from a hypothesis to itself) must be all 0s}

\item{names}{(Optional) A character vector of hypothesis names. If not
provided, names from \code{hypotheses} and \code{transitions} will be used. If names
are not specified, hypotheses will be named sequentially as H1, H2, ...}
}
\value{
An S3 object of class \code{initial_graph}. The underlying structure is a
list with elements \code{hypotheses} and \code{transitions}
}
\description{
A multiple comparison procedure graph can be represented by 1) a vector of
initial hypothesis weights, and 2) a matrix of initial transition weights.
This function also does validation to make sure that the vector and matrix
combine to form a valid MCP graph.
}
\examples{
# A graphical multiple comparison procedure with two primary hypotheses (H1
# and H2) and two secondary hypotheses (H3 and H4)
# See Figure 1 in Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer,
# W., & Rohmeyer, K. (2011). Graphical approaches for multiple comparison
# procedures using weighted Bonferroni, Simes, or parametric tests.
# Biometrical Journal, 53(6), 894-913.
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
g

# Explicit names override names in `hypotheses` (with a warning)
hypotheses <- c(h1 = 0.5, h2 = 0.5, h3 = 0, h4 = 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
g

# Explicit names override names in `transitions` (with a warning)
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  h1 = c(0, 0, 1, 0),
  h2 = c(0, 0, 0, 1),
  h3 = c(0, 1, 0, 0),
  h4 = c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
g

# Use names in `hypotheses`
hypotheses <- c(H1 = 0.5, H2 = 0.5, H3 = 0, H4 = 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)
g

# Use names in `transitions`
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  H1 = c(0, 0, 1, 0),
  H2 = c(0, 0, 0, 1),
  H3 = c(0, 1, 0, 0),
  H4 = c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)
g

# When names are not specified, hypotheses are numbered sequentially as
# H1, H2, ...
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)
g
}
