% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_critical.R
\name{c_function}
\alias{c_function}
\alias{solve_c}
\alias{calculate_critical_parametric}
\alias{calculate_critical_simes}
\title{Calculate updated hypothesis weights for the closure of a graph}
\usage{
c_function(x, weights, corr, alpha)

solve_c(weights, corr, alpha)

calculate_critical_parametric(intersections, corr, alpha, groups)

calculate_critical_simes(intersections, p, groups)
}
\arguments{
\item{x}{The root to solve for with \code{\link[stats:uniroot]{stats::uniroot()}}}

\item{weights}{A numeric vector of hypothesis weights}

\item{corr}{A numeric matrix of correlations between hypotheses' test
statistics}

\item{alpha}{A numeric scalar specifying the global significance level for
testing}

\item{intersections}{For parametric, a compact representation of
\code{\link[=generate_weights]{generate_weights()}} output, where missing hypotheses get a missing value
for weights, and h-vectors are dropped. For Simes, just the weights from
\code{\link[=generate_weights]{generate_weights()}} output}

\item{groups}{A list of numeric vectors specifying hypotheses to test
together}

\item{p}{A numeric vector of p-values}
}
\value{
Outputs:
\itemize{
\item For \verb{calculate_critical_*()}, a matrix with the same shape as
\code{intersections}, where the weights have been adjusted according to the
specified adjustment method
\item For \code{c_function()}, the critical value for the given group
}
}
\description{
The weights created by \code{\link[=generate_weights]{generate_weights()}} work immediately for Bonferroni
testing, but parametric and Simes testing require additional calculations.
The \verb{calculate_critical_*()} functions apply parametric or Simes weight
increases to get updated weights for testing. They also subset the weights
columns by the appropriate groups
}
\examples{
p <- 1:6 / 200

g <- bonferroni_holm(6)
gw_large <- generate_weights(g)

gw_0 <- gw_large[, 7:12]
gw <- ifelse(gw_large[, 1:6], gw_0, NA)

graphicalMCP:::calculate_critical_parametric(gw, diag(6), .05, list(1:3))
graphicalMCP:::calculate_critical_simes(gw_0, p, list(4:6))
}
\keyword{(2009).}
\keyword{(2011).}
\keyword{(2016).}
\keyword{(2017).}
\keyword{(2019).}
\keyword{1489–1501.}
\keyword{28(4),}
\keyword{30(13),}
\keyword{35(22),}
\keyword{38(27),}
\keyword{4041–4055.}
\keyword{5268–5282.}
\keyword{53(6),}
\keyword{586–604.}
\keyword{59(5),}
\keyword{894–913.}
\keyword{918–931.}
\keyword{A}
\keyword{Biometrical}
\keyword{Bonferroni}
\keyword{Bonferroni,}
\keyword{Brannath,}
\keyword{Bretz,}
\keyword{D.,}
\keyword{E.,}
\keyword{F.}
\keyword{F.,}
\keyword{G.}
\keyword{Glimm,}
\keyword{Graphical}
\keyword{Hochberg}
\keyword{Hommel,}
\keyword{Journal,}
\keyword{K.}
\keyword{Klinglmueller,}
\keyword{Lu,}
\keyword{M.}
\keyword{M.,}
\keyword{Maurer,}
\keyword{Medicine,}
\keyword{Posch,}
\keyword{Rohmeyer,}
\keyword{Simes}
\keyword{Simes,}
\keyword{Statistics}
\keyword{Symmetric}
\keyword{Test}
\keyword{W.,}
\keyword{Xi,}
\keyword{a}
\keyword{analyses}
\keyword{and}
\keyword{application}
\keyword{approach}
\keyword{approaches}
\keyword{comparison}
\keyword{considerations}
\keyword{endpoint}
\keyword{equally}
\keyword{for}
\keyword{framework}
\keyword{graphical}
\keyword{graphs}
\keyword{https://doi.org/10.1002/bimj.201000239}
\keyword{https://doi.org/10.1002/bimj.201600233}
\keyword{https://doi.org/10.1002/sim.3495}
\keyword{https://doi.org/10.1002/sim.3988}
\keyword{https://doi.org/10.1002/sim.6985}
\keyword{https://doi.org/10.1002/sim.8375}
\keyword{in}
\keyword{internal}
\keyword{mixture}
\keyword{multiple}
\keyword{of}
\keyword{or}
\keyword{parametric}
\keyword{power}
\keyword{procedure.}
\keyword{procedures}
\keyword{procedures.}
\keyword{rejective}
\keyword{sequentially}
\keyword{test}
\keyword{tests,}
\keyword{tests.}
\keyword{the}
\keyword{to}
\keyword{unified}
\keyword{using}
\keyword{weighted}
\keyword{with}
