% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_graphs.R
\name{bonferroni_holm}
\alias{bonferroni_holm}
\alias{huque_alosh_bhore_2011}
\alias{wiens_dmitrienko_2005}
\alias{fixed_sequence}
\alias{fallback}
\alias{simple_successive_1}
\alias{simple_successive_2}
\alias{random_graph}
\alias{complex_example_1}
\alias{complex_example_2}
\title{Example graphs from a selection of papers}
\usage{
bonferroni_holm(num_hyps, hyp_names = NULL)

bonferroni_holm(num_hyps, hyp_names = NULL)

huque_alosh_bhore_2011(hyp_names = NULL)

wiens_dmitrienko_2005(hyp_names = NULL)

fixed_sequence(num_hyps = 3, hyp_names = NULL)

fallback(hypotheses = c(1, 0, 0), hyp_names = NULL)

simple_successive_1(hyp_names = NULL)

simple_successive_2(hyp_names = NULL)

random_graph(num_hyps, hyp_names = NULL)

complex_example_1(hyp_names = NULL)

complex_example_2(hyp_names = NULL)

complex_example_2(hyp_names = NULL)
}
\arguments{
\item{num_hyps}{Number of vertices in the graph}

\item{hyp_names}{(Optional) A character vector of hypothesis names. If names
are not specified, hypotheses will be named sequentially as H1, H2, ...}

\item{hypotheses}{Hypothesis weights in a fallback procedure}
}
\value{
An S3 object as returned by \code{\link[=graph_create]{graph_create()}}
}
\description{
Example graphs from a selection of papers
}
\examples{
bonferroni_holm(3, hyp_names = paste("dose", letters[1:3]))
}
\references{
Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical
approach to sequentially rejective multiple test procedures. Statistics in
Medicine, 28(4), 586–604. \url{https://doi.org/10.1002/sim.3495}

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations
for multiple endpoint analyses using sequentially rejective graphical
procedures. Statistics in Medicine, 30(13), 1489–1501.
\url{https://doi.org/10.1002/sim.3988}

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer,
K. (2011). Graphical approaches for multiple comparison procedures using
weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6),
894–913. \url{https://doi.org/10.1002/bimj.201000239}

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted
Simes tests. Statistics in Medicine, 35(22), 4041–4055.
\url{https://doi.org/10.1002/sim.6985}

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for
weighted parametric multiple test procedures. Biometrical Journal, 59(5),
918–931. \url{https://doi.org/10.1002/bimj.201600233}

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests,
with application to the Hochberg procedure. Statistics in Medicine, 38(27),
5268–5282. \url{https://doi.org/10.1002/sim.8375}

Rohmeyer K, Klinglmueller F (2020). \emph{gMCP: Graph Based Multiple Test
Procedures}. R package version 0.8-15,
\url{https://cran.r-project.org/package=gMCP}.
}
