% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{generate_weights}
\alias{generate_weights}
\title{Generate weights for the full closure tree of an MCP graph}
\usage{
generate_weights(graph)
}
\arguments{
\item{graph}{An MCP graph as created by \code{create_graph()}}
}
\value{
A numeric matrix of all subgraph weights. The first half of columns
indicate which hypotheses are included in the given subgraph, and the
second half of columns are the weights
}
\description{
Generate weights for the full closure tree of an MCP graph
}
\section{Performance}{

An evaluation of a few different methods to generate the weights of all
subgraphs according to Bretz et al. 2011 can be found in the perf-tests
directory. On a Workbench session with sufficient RAM (Not sure how much
CPU), the \code{gMCP::generateWeights()} method runs significantly faster for the
size 2 case, but it grows quickly in both computing time and memory
}

\examples{

ex_graph <- create_graph(
  hypotheses = c(.5, .5, 0, 0),
  transitions = rbind(
    c(0, 0, 1, 0),
    c(0, 0, 0, 1),
    c(0, 1, 0, 0),
    c(1, 0, 0, 0)
  )
)

generate_weights(ex_graph)

}
