% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generate_weights.R
\name{graph_generate_weights}
\alias{graph_generate_weights}
\title{Calculate weights for the closure of a graph}
\usage{
graph_generate_weights(graph)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=graph_create]{graph_create()}}}
}
\value{
A numeric matrix of all intersection hypothesis weights. Each row
corresponds to a single intersection hypothesis. The first half of the
columns indicate which hypotheses are included in the given intersection
hypothesis, and the second half of columns are the weights
}
\description{
The closure of a graph is the set of all sub-graphs, or intersections
hypotheses, of a graph. Weights for each sub-graph are calculated using the
weighting strategy defined in Bretz et al (2011).
}
\section{Performance}{

Much thought was given to the performance of this code, as the memory and
time usage can grow quickly as graph size grows. On the systems used for
testing, a size 10 graph had a median run time of 20-60 ms. Run time
increases at a rate of O(2 ^ n), so e.g. a size 5 graph takes approximately
twice as long to run as a size 4 graph. See
\code{vignette("generate-weights-performance")} for a detailed analysis and
explanation
}

\examples{

par_gate <- graph_create(
  hypotheses = c(.5, .5, 0, 0),
  transitions = rbind(
    c(0, 0, 1, 0),
    c(0, 0, 0, 1),
    c(0, 1, 0, 0),
    c(1, 0, 0, 0)
  )
)

graph_generate_weights(par_gate)

}
\references{
Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical
approach to sequentially rejective multiple test procedures. Statistics in
Medicine, 28(4), 586b

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations
for multiple endpoint analyses using sequentially rejective graphical
procedures. Statistics in Medicine, 30(13), 1489b
\url{https://doi.org/10.1002/sim.3988}

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer,
K. (2011). Graphical approaches for multiple comparison procedures using
weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6),
894b

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted
Simes tests. Statistics in Medicine, 35(22), 4041b
\url{https://doi.org/10.1002/sim.6985}

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for
weighted parametric multiple test procedures. Biometrical Journal, 59(5),
918b

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests,
with application to the Hochberg procedure. Statistics in Medicine, 38(27),
5268b

Rohmeyer K, Klinglmueller F (2020). \emph{gMCP: Graph Based Multiple Test
Procedures}. R package version 0.8-15,
\url{https://cran.r-project.org/package=gMCP}.
}
