% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_rejection_orderings.R
\name{graph_rejection_orderings}
\alias{graph_rejection_orderings}
\title{Find alternate rejection orderings for shortcut testing}
\usage{
graph_rejection_orderings(shortcut_test_result)
}
\arguments{
\item{shortcut_test_result}{A \code{graph_report} object as returned by
\code{\link[=graph_test_shortcut]{graph_test_shortcut()}}}
}
\value{
A modified \code{graph_report} object containing all valid orderings for
deleting the significant hypotheses
}
\description{
When using \code{graph_test_shortcut()}, there may be multiple hypotheses which
can be rejected at each step. The specific hypothesis chosen is decided based
on the minimum adjusted p-value, but final results don't change if rejection
order changes. This function shows every other order that deletes the same
hypotheses, where each hypothesis is still a valid rejection at the step it
is deleted.
}
\examples{
graph <- simple_successive_2()

short_res <- graph_test_shortcut(graph, c(.018, .01, .03, .004))

# Reject H1, H2, and H4
short_res$outputs$rejected

# But these cannot be rejected in any order - For instance, H4 has 0 weight
# in the initial graph and cannot be rejected first
graph_rejection_orderings(short_res)

# Finally, intermediate graphs can be obtained by putting one of the
# orderings outputs into `graph_update()`
graph_update(graph, delete = c(2, 1, 4))
}
