% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalMCP-package.R
\docType{package}
\name{graphicalMCP-package}
\alias{graphicalMCP}
\alias{graphicalMCP-package}
\title{graphicalMCP: Graphical Approach for Multiple Comparison Procedures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A multiple comparison procedure (or multiple test procedure) is a a statistical analysis method for determining efficacy of multiple drugs, or multiple doses of the same drug, in a single clinical trial. (Bretz et al., 2011) laid out a graph-based approach to these multiple comparison procedures, in which the weights of the vertices and edges of the graph are determined independently of the particular statistical test used to assess results. This is a low-dependency implementation of the methods described in that and subsequent papers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://urban-sniffle-p11zlpj.pages.github.io/}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/graphicalMCP/issues}
}

}
\author{
\strong{Maintainer}: Dong Xi \email{dong.xi1@gilead.com}

Authors:
\itemize{
  \item Ethan Brockmann \email{ethan.brockmann@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences, Inc. [copyright holder, funder]
}

}
\keyword{internal}
