% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust_bonferroni}
\alias{p_adjust_bonferroni}
\alias{p_adjust_parametric}
\alias{p_adjust_simes}
\alias{p_adjust}
\title{Calculate adjusted p-values}
\usage{
p_adjust_bonferroni(p_values, weights, corr = NULL)

p_adjust_parametric(p_values, weights, corr = NULL)

p_adjust_simes(p_values, weights, corr = NULL)

p_adjust(p_values, weights, groups, tests, corr = NULL)
}
\arguments{
\item{p_values}{A numeric vector of p-values}

\item{weights}{A named numeric vector of weights}

\item{corr}{(Optional) A correlation matrix to use in parametric testing}

\item{groups}{A list of groups of hypotheses to test together}

\item{tests}{A character vector of test names}
}
\value{
A single adjusted p-value for the given group

A vector of adjusted p-values, one for each group
}
\description{
This basically does the \verb{do.call(p_adjust_<test name>, ...)} portion of
\code{test_graph()}. The scope of each function is a little fuzzier in this older
version
}
\examples{
set.seed(22723)

w <- c("H1" = .75, "H2" = .25, "H3" = 0)
p <- c("H1" = .019, "H2" = .025, "H3" = .05)

p_adjust_bonferroni(p, w)
p_adjust_simes(p, w)

corr1 <- diag(3)
corr2 <- corr1
corr2[lower.tri(corr2)] <- corr2[upper.tri(corr2)] <- runif(3, -1, 1)

# No correlation
p_adjust_parametric(p, w, corr1)

# Uniform random pairwise correlations
p_adjust_parametric(p, w, corr2)
}
