% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_p.R
\name{p_adjust}
\alias{p_adjust}
\title{Calculate adjusted P-values}
\usage{
p_adjust(p_values, weights, groups, tests, corr = NULL)
}
\arguments{
\item{corr}{}
}
\value{
Numeric vector with one entry per testing group. Can be thought of as
p_J_h in a variety of papers, where J is the current intersection, and h is
a group
}
\description{
Calculate adjusted P-values
}
