% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust_bonferroni}
\alias{p_adjust_bonferroni}
\alias{p_adjust_parametric}
\alias{p_adjust_simes}
\title{Calculate adjusted p-values}
\usage{
p_adjust_bonferroni(p, weights)

p_adjust_parametric(p, weights, corr = NULL)

p_adjust_simes(p, weights)
}
\arguments{
\item{p}{A named numeric vector of p-values}

\item{weights}{A named numeric vector of hypothesis weights}

\item{corr}{(Optional) A numeric matrix of correlations between hypotheses'
test statistics}
}
\value{
A single adjusted p-value for the given group
}
\description{
Calculate adjusted p-values
}
\examples{
set.seed(22723)

w <- c("H1" = .75, "H2" = .25, "H3" = 0)
p <- c("H1" = .019, "H2" = .025, "H3" = .05)

graphicalMCP:::p_adjust_bonferroni(p, w)
graphicalMCP:::p_adjust_simes(p, w)

corr1 <- diag(3)
corr2 <- corr1
corr2[lower.tri(corr2)] <- corr2[upper.tri(corr2)] <- runif(3, -1, 1)

# No correlation
graphicalMCP:::p_adjust_parametric(p, w, corr1)

# Uniform random pairwise correlations
graphicalMCP:::p_adjust_parametric(p, w, corr2)
}
\keyword{(2009).}
\keyword{(2011).}
\keyword{(2016).}
\keyword{(2017).}
\keyword{(2019).}
\keyword{1489–1501.}
\keyword{28(4),}
\keyword{30(13),}
\keyword{35(22),}
\keyword{38(27),}
\keyword{4041–4055.}
\keyword{5268–5282.}
\keyword{53(6),}
\keyword{586–604.}
\keyword{59(5),}
\keyword{894–913.}
\keyword{918–931.}
\keyword{A}
\keyword{Biometrical}
\keyword{Bonferroni}
\keyword{Bonferroni,}
\keyword{Brannath,}
\keyword{Bretz,}
\keyword{D.,}
\keyword{E.,}
\keyword{F.}
\keyword{F.,}
\keyword{G.}
\keyword{Glimm,}
\keyword{Graphical}
\keyword{Hochberg}
\keyword{Hommel,}
\keyword{Journal,}
\keyword{K.}
\keyword{Klinglmueller,}
\keyword{Lu,}
\keyword{M.}
\keyword{M.,}
\keyword{Maurer,}
\keyword{Medicine,}
\keyword{Posch,}
\keyword{Rohmeyer,}
\keyword{Simes}
\keyword{Simes,}
\keyword{Statistics}
\keyword{Symmetric}
\keyword{Test}
\keyword{W.,}
\keyword{Xi,}
\keyword{a}
\keyword{analyses}
\keyword{and}
\keyword{application}
\keyword{approach}
\keyword{approaches}
\keyword{comparison}
\keyword{considerations}
\keyword{endpoint}
\keyword{equally}
\keyword{for}
\keyword{framework}
\keyword{graphical}
\keyword{graphs}
\keyword{https://doi.org/10.1002/bimj.201000239}
\keyword{https://doi.org/10.1002/bimj.201600233}
\keyword{https://doi.org/10.1002/sim.3495}
\keyword{https://doi.org/10.1002/sim.3988}
\keyword{https://doi.org/10.1002/sim.6985}
\keyword{https://doi.org/10.1002/sim.8375}
\keyword{in}
\keyword{internal}
\keyword{mixture}
\keyword{multiple}
\keyword{of}
\keyword{or}
\keyword{parametric}
\keyword{power}
\keyword{procedure.}
\keyword{procedures}
\keyword{procedures.}
\keyword{rejective}
\keyword{sequentially}
\keyword{test}
\keyword{tests,}
\keyword{tests.}
\keyword{the}
\keyword{to}
\keyword{unified}
\keyword{using}
\keyword{weighted}
\keyword{with}
