% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust.R
\name{p_adjust_bonferroni}
\alias{p_adjust_bonferroni}
\alias{p_adjust_parametric}
\alias{p_adjust_simes}
\title{Calculate adjusted p-values}
\usage{
p_adjust_bonferroni(p_values, weights, corr = NULL)

p_adjust_parametric(p_values, weights, corr = NULL)

p_adjust_simes(p_values, weights, corr = NULL)
}
\arguments{
\item{p_values}{A named numeric vector of p-values to adjust}

\item{weights}{A named numeric vector of weights to adjust the p-values by}

\item{corr}{(Optional) A numeric matrix indicating the correlation between
the test statistics which generated the p-values. For parametric testing,
\code{corr} must be a square matrix with side length equal to the length of \code{p}
and \code{weights}. Ignored for Simes and Bonferroni testing}
}
\value{
A single adjusted p-value for the given group
}
\description{
Calculate adjusted p-values
}
\examples{
set.seed(22723)

w <- c("H1" = .75, "H2" = .25, "H3" = 0)
p <- c("H1" = .019, "H2" = .025, "H3" = .05)

graphicalMCP:::p_adjust_bonferroni(p, w)
graphicalMCP:::p_adjust_simes(p, w)

corr1 <- diag(3)
corr2 <- corr1
corr2[lower.tri(corr2)] <- corr2[upper.tri(corr2)] <- runif(3, -1, 1)

# No correlation
graphicalMCP:::p_adjust_parametric(p, w, corr1)

# Uniform random pairwise correlations
graphicalMCP:::p_adjust_parametric(p, w, corr2)
}
