% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.initial_graph.R
\name{plot.initial_graph}
\alias{plot.initial_graph}
\title{S3 plot method for the class \code{initial_graph}}
\usage{
\method{plot}{initial_graph}(
  graph,
  layout = igraph::layout_nicely,
  groups = list(graph$hypotheses),
  groups_layout = NULL,
  nrow = NULL,
  ncol = NULL,
  edge_curves = NULL,
  precision = 4,
  eps = 0.001,
  background_color = "white",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
S3 plot method for the class \code{initial_graph}
}
