% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.updated_graph.R
\name{plot.updated_graph}
\alias{plot.updated_graph}
\title{S3 plot method for the class \code{updated_graph}}
\usage{
\method{plot}{updated_graph}(x, ...)
}
\arguments{
\item{x}{An \code{updated_graph} object as returned by \code{graph_update()}}

\item{...}{
  Arguments passed on to \code{\link[=plot.initial_graph]{plot.initial_graph}}
  \describe{
    \item{\code{layout}}{An igraph layout specification (See \code{?igraph.plotting}), or
\code{"grid"}, which lays out hypotheses left-to-right, top-to-bottom with
\code{nrow} rows and \code{ncol} columns.}
    \item{\code{nrow}}{An integer scalar specifying the number of rows in the vertex
grid. If row and columns counts are not specified, vertices will be laid
out as close to a square as possible.}
    \item{\code{ncol}}{An integer scalar specifying the number of columns in the vertex
grid. If row and columns counts are not specified, vertices will be laid
out as close to a square as possible.}
    \item{\code{edge_curves}}{A named numeric vector specifying the curvature edges.
Edge pairs (Where two vertices share an edge in each possible direction)
are detected automatically and get 0.25 curvature. Adjust edges by adding
an entry with name \verb{"vertex1|vertex2}. Adjust default edge pairs curvature
by adding an entry with name \code{"pairs"}.}
    \item{\code{precision}}{An integer scalar specifying how many decimal places should
be displayed for weights}
    \item{\code{eps}}{A numeric scalar. Edge weights between 0 and \code{eps} will be
displayed as \eqn{\epsilon}, and edge weights between \code{1 - eps} and 1 will
be displayed as \eqn{1 - \epsilon}}
    \item{\code{background_color}}{A character scalar specifying a background color for
the whole plotting area. Passed directly to \code{\link[graphics:par]{graphics::par()}} (\code{bg})}
    \item{\code{margins}}{A length 4 numeric vector specifying the margins for the plot.
Defaults to all 0, since igraph plots tend to have large margins. It is
passed directly to \code{\link[graphics:par]{graphics::par()}} (\code{mar})}
  }}
}
\value{
NULL, after plotting the graph
}
\description{
S3 plot method for the class \code{updated_graph}
}
\examples{
plot(
  graph_update(
    simple_successive_2(),
    c(TRUE, FALSE, TRUE, TRUE)
   ),
  layout = "grid"
)
}
