% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.graph_report.R
\name{print.graph_report}
\alias{print.graph_report}
\title{S3 print method for class \code{graph_report}}
\usage{
\method{print}{graph_report}(x, ..., precision = 4, indent = 2, rows = 10)
}
\arguments{
\item{x}{An object of class \code{graph_report} to print}

\item{...}{Other values passed on to other methods (currently unused)}

\item{precision}{An integer scalar indicating the number of significant
figures to include in numeric values}

\item{indent}{An integer scalar indicating how many spaces to indent results}

\item{rows}{An integer scalar indicating how many rows of verbose and test
values output to print}
}
\description{
A graph report displays
\itemize{
\item The initial graph being tested,
\item p-values & significance level used for tests,
\item Which hypotheses can be rejected, and
\item Optional detailed test results, which vary depending on whether shortcut or
closed testing was used
}
}
