% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.initial_graph.R
\name{print.initial_graph}
\alias{print.initial_graph}
\title{S3 print method for the class \code{initial_graph}}
\usage{
\method{print}{initial_graph}(x, ..., precision = 4, indent = 0, title = "Initial graph")
}
\arguments{
\item{x}{An object of class \code{initial_graph} to print}

\item{...}{Other values passed on to other methods (currently unused)}

\item{precision}{An integer scalar indicating the number of significant figures
to include in numeric values}

\item{indent}{An integer scalar indicating how many spaces to indent results}

\item{title}{First line of printing, used internally to distinguish initial
graphs from updated graphs}
}
\description{
A printed \code{initial_graph} displays a header stating what the object is, the
hypothesis weights, and the transition weights.
}
