% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_report_print.R
\name{print.power_report}
\alias{print.power_report}
\title{S3 print method for class \code{power_report}}
\usage{
\method{print}{power_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{power_report} to print}

\item{...}{Other values passed on to other methods (currently unused)}
}
\description{
A power report displays
\itemize{
\item The initial graph being tested,
\item Testing and simulation options,
\item Detailed p-values and test results, and
\item Final power calculations
}
}
