% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.updated_graph.R
\name{print.updated_graph}
\alias{print.updated_graph}
\title{S3 print method for class \code{updated_graph}}
\usage{
\method{print}{updated_graph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{updated_graph} to print}

\item{...}{Other values passed on to other methods (currently unused)}
}
\description{
A printed \code{updated_graph} displays the initial graph, the vector of
keep/delete hypotheses, and the updated graph
}
