% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_graph.R
\name{test_graph}
\alias{test_graph}
\title{Apply weighted Bonferroni, parametric, and Simes tests}
\usage{
test_graph(
  graph,
  p_values,
  alpha = 0.05,
  tests = list(bonferroni = list(seq_along(graph$hypotheses)), parametric = NULL, simes =
    NULL),
  corr = NULL,
  use_cj = FALSE
)
}
\arguments{
\item{graph}{An initial graph as created by \code{create_graph()}}

\item{p_values}{A numeric vector of p-values, one for each hypothesis in the
graph}

\item{alpha}{A numeric vector of length 1 specifying the un-weighted alpha
level at which to test each hypothesis}

\item{tests}{A list with three elements, \code{bonferroni}, \code{simes}, and
\code{parametric}. Each element is a list of hypothesis groups to apply the
given test to. Each hypothesis must be specified exactly once, so that the
length of all elements of the list equals the number of hypotheses. The
default is to apply the weighted Bonferroni test to all hypotheses}

\item{corr}{(Optional) A correlation matrix for the test statistics of
\code{graph}. Diagonal entries should be 1. A known absence of correlation
should be entered as 0, and unknown correlation should be entered as NA.
This argument is only used for the parametric test. For each element of a
parametric testing group, the correlation matrix for that group's sub-graph
must be fully known if \code{use_cj} is FALSE. If \code{use_cj} is TRUE, the whole
correlation matrix must be known}

\item{use_cj}{A logical vector of length one. If FALSE (the default), then
the critical value \code{c} is re-calculated for each parametric testing group
within each sub-graph. If TRUE, then \code{c} is calculated once for each
intersection hypothesis}
}
\value{
A \code{graph_report} object, consisting of
\itemize{
\item The initial graph being tested,
\item p-values & alpha used for tests,
\item Which hypotheses can be rejected, and
\item Detailed test results matrix, including the results of
\code{generate_weights()} & test results for each intersection hypothesis
}
}
\description{
Apply weighted Bonferroni, parametric, and Simes tests
}
\examples{

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- create_graph(hypotheses, transitions)

corr <- matrix(nrow = 4, ncol = 4)
corr[3:4, 3:4] <- .5
diag(corr) <- 1

test_graph(
  g,
  p_values = c(.01, .02, .05, .1),
  alpha = .05,
  tests = list(
    bonferroni = 1,
    simes = list(c(2)),
    parametric = list(c(3, 4))
  ),
  corr = corr
)
}
