% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_tests.R
\name{test_graph_fast}
\alias{test_graph_fast}
\alias{test_graph_shortcut_cpp}
\title{Calculate hypothesis rejection results efficiently}
\usage{
test_graph_fast(p, alpha, critical_values, intersections)

test_graph_shortcut_cpp(graph, p, alpha = 0.025)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{alpha}{A numeric scalar specifying the global significance level for
testing}

\item{critical_values}{The weights (second half of columns) from
\code{\link[=generate_weights]{generate_weights()}} output, adjusted by the appropriate testing algorithm
(Bonferroni, Simes, or parametric)}

\item{intersections}{The first half of columns from \code{\link[=generate_weights]{generate_weights()}}
output, indicating which hypotheses are contained in each intersection}

\item{graph}{An initial graph as returned by \code{\link[=create_graph]{create_graph()}}}
}
\value{
A logical or integer vector of results indicating whether each
hypothesis can be accepted or rejected globally.
}
\description{
For insight and nice reporting, prefer \code{\link[=test_graph_closure]{test_graph_closure()}} or
\code{\link[=test_graph_shortcut]{test_graph_shortcut()}} with all of their options. They are reasonably fast
for interactive use. However in order to minimize power run time, more
efficient testing functions are available. The power simulation can be
segmented so that certain parts, like generating weights and calculating some
critical values, can be done only a single time. The closure testing function
has been stripped down to just a few vectorized lines for efficiency. A
separate optimized function is available for testing a graph with the
Bonferroni sequential shortcut.
}
\examples{
par_gate <- simple_successive_1()
num_hyps <- length(par_gate$hypotheses)

p <- c(.001, .02, .002, .03)

weighting_strategy <- generate_weights(par_gate)
intersections <- weighting_strategy[, seq_len(num_hyps), drop = FALSE]

graphicalMCP:::test_graph_fast(p, .025, weighting_strategy, intersections)
graphicalMCP:::test_graph_shortcut_cpp(par_gate, p, .025)
}
\references{
Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical
approach to sequentially rejective multiple test procedures. Statistics in
Medicine, 28(4), 586–604. \url{https://doi.org/10.1002/sim.3495}

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations
for multiple endpoint analyses using sequentially rejective graphical
procedures. Statistics in Medicine, 30(13), 1489–1501.
\url{https://doi.org/10.1002/sim.3988}

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer,
K. (2011). Graphical approaches for multiple comparison procedures using
weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6),
894–913. \url{https://doi.org/10.1002/bimj.201000239}

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted
Simes tests. Statistics in Medicine, 35(22), 4041–4055.
\url{https://doi.org/10.1002/sim.6985}

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for
weighted parametric multiple test procedures. Biometrical Journal, 59(5),
918–931. \url{https://doi.org/10.1002/bimj.201600233}

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests,
with application to the Hochberg procedure. Statistics in Medicine, 38(27),
5268–5282. \url{https://doi.org/10.1002/sim.8375}

Rohmeyer K, Klinglmueller F (2020). \emph{gMCP: Graph Based Multiple Test
Procedures}. R package version 0.8-15,
\url{https://cran.r-project.org/package=gMCP}.
}
\seealso{
\code{\link[=test_graph_closure]{test_graph_closure()}}, \code{\link[=test_graph_shortcut]{test_graph_shortcut()}}
}
\keyword{internal}
