% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonferroni_sequential.R, R/graph_report.R,
%   R/test_power_input_val.R
\name{bonferroni_sequential}
\alias{bonferroni_sequential}
\alias{bs_cpp}
\alias{test_graph}
\alias{test_input_val}
\title{Test a graph}
\usage{
bonferroni_sequential(graph, p, alpha = 0.05, critical = FALSE)

bs_cpp(graph, p, alpha = 0.05)

test_graph(
  graph,
  p,
  alpha = 0.05,
  groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  corr = NULL,
  verbose = FALSE,
  critical = FALSE
)

test_input_val(
  graph,
  p,
  alpha,
  groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  corr,
  verbose,
  critical
)
}
\arguments{
\item{graph}{An initial graph as returned by \code{create_graph()}}

\item{p}{A numeric vector of p-values}

\item{alpha}{A numeric scalar specifying the global level to test at}

\item{critical}{A logical scalar specifying whether hypothesis-level detail
should be included in the results, including calculating critical values
for parametric tests}

\item{groups}{A list of numeric vectors specifying hypotheses to test
together}

\item{test_types}{A character vector of tests to apply to the given groups}

\item{corr}{(Optional) A numeric matrix of correlations between hypotheses'
test statistics}

\item{verbose}{A logical scalar specifying whether the results for each
intersection hypothesis should be included}
}
\value{
A \code{graph_report} object, a list of 4 elements: \code{inputs}, \code{outputs},
\code{verbose}, and \code{critical}
\itemize{
\item Inputs - A list of the input parameters used to run the test
\item Outputs - A list of global test results
\item Verbose - A matrix with detailed adjusted p-value results
\item Critical - A data frame with hypothesis-level test details for each
intersection
}
}
\description{
Test a graph
}
\examples{

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- create_graph(hypotheses, transitions)
p <- c(.01, .005, .015, .022)

corr <- matrix(nrow = 4, ncol = 4)
corr[3:4, 3:4] <- .9
diag(corr) <- 1

corr2 <- matrix(.5, nrow = 4, ncol = 4)
diag(corr2) <- 1

# The default is all Bonferroni with alpha = .05
test_graph(g, p)

# But tests can be specified at the hypothesis-level
test_graph(
  graph = g,
  p = p,
  alpha = .025,
  groups = list(1, 2, 3:4),
  test_types = c("bonferroni", "simes", "parametric"),
  corr = corr
)
}
