% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_graph_closure.R, R/test_graph_shortcut.R
\name{test_graph_closure}
\alias{test_graph_closure}
\alias{test_graph_shortcut}
\title{Report details of hypothesis rejections}
\usage{
test_graph_closure(
  graph,
  p,
  alpha = 0.025,
  groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  corr = NULL,
  verbose = FALSE,
  critical = FALSE
)

test_graph_shortcut(graph, p, alpha = 0.025, verbose = FALSE, critical = FALSE)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=create_graph]{create_graph()}}}

\item{p}{A numeric vector of p-values}

\item{alpha}{A numeric scalar specifying the global significance level for
testing}

\item{groups}{A list of numeric vectors specifying hypotheses to test
together}

\item{test_types}{A character vector of tests to apply to the given groups}

\item{corr}{(Optional) A numeric matrix of correlations between hypotheses'
test statistics}

\item{verbose}{A logical scalar specifying whether the results for each
intersection hypothesis should be included}

\item{critical}{A logical scalar specifying whether hypothesis-level detail
should be included in the results, including calculating critical values
for parametric tests}
}
\value{
A \code{graph_report} object, a list of 4 elements: \code{inputs}, \code{outputs},
\code{verbose}, and \code{critical}
\itemize{
\item Inputs - A list of the input parameters used to run the test
\item Outputs - A list of global test results
\item Verbose - A matrix with detailed adjusted p-value results (graph deletion
sequence for shortcut testing)
\item Critical - A data frame with hypothesis-level test details for each
intersection (each step for shortcut testing)
}
}
\description{
The slower graph testing functions have design choices made that favor ease
of interpreting results over speed. Results include hypothesis rejection
decisions, but also the test values that led to the final result.
The functions include options for reporting details using the adjusted
p-value method or critical value method.
}
\examples{

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- create_graph(hypotheses, transitions)
p <- c(.01, .005, .015, .022)

corr <- matrix(nrow = 4, ncol = 4)
corr[3:4, 3:4] <- .9
diag(corr) <- 1

corr2 <- matrix(.5, nrow = 4, ncol = 4)
diag(corr2) <- 1

# The default is all Bonferroni with alpha = .025
test_graph_closure(g, p)

# But tests can be specified at the hypothesis-level
test_graph_closure(
  graph = g,
  p = p,
  alpha = .025,
  groups = list(1, 2, 3:4),
  test_types = c("bonferroni", "simes", "parametric"),
  corr = corr
)
}
\references{
Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical
approach to sequentially rejective multiple test procedures. Statistics in
Medicine, 28(4), 586–604. \url{https://doi.org/10.1002/sim.3495}

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations
for multiple endpoint analyses using sequentially rejective graphical
procedures. Statistics in Medicine, 30(13), 1489–1501.
\url{https://doi.org/10.1002/sim.3988}

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer,
K. (2011). Graphical approaches for multiple comparison procedures using
weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6),
894–913. \url{https://doi.org/10.1002/bimj.201000239}

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted
Simes tests. Statistics in Medicine, 35(22), 4041–4055.
\url{https://doi.org/10.1002/sim.6985}

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for
weighted parametric multiple test procedures. Biometrical Journal, 59(5),
918–931. \url{https://doi.org/10.1002/bimj.201600233}

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests,
with application to the Hochberg procedure. Statistics in Medicine, 38(27),
5268–5282. \url{https://doi.org/10.1002/sim.8375}

Rohmeyer K, Klinglmueller F (2020). \emph{gMCP: Graph Based Multiple Test
Procedures}. R package version 0.8-15,
\url{https://cran.r-project.org/package=gMCP}.
}
\seealso{
\code{\link[=test_graph_fast]{test_graph_fast()}}, \code{\link[=test_graph_shortcut_cpp]{test_graph_shortcut_cpp()}}
}
