% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_graph.R
\name{update_graph}
\alias{update_graph}
\title{Delete hypotheses from a graph}
\usage{
update_graph(graph, keep)
}
\arguments{
\item{graph}{An initial graph as returned by \code{\link[=create_graph]{create_graph()}}}

\item{keep}{A logical or integer vector, denoting which hypotheses to
keep/delete. An entry of \code{FALSE} (or 0) corresponds to a deletion, and
\code{TRUE} (or 1) corresponds to keeping a hypothesis}
}
\value{
An object of class \code{updated_graph} with 3 elements
\itemize{
\item The initial graph object
\item The Boolean vector indicating which hypotheses are kept/deleted
\item The updated graph object with specific hypotheses deleted
}
}
\description{
It is not always obvious what a graph's weights will look like after deleting
one or more hypotheses. While \code{\link[=generate_weights]{generate_weights()}} calculates all sub-graphs'
hypothesis weights, \code{update_graph()} gives a more detailed view for a single
set of deletions, including transition weights as well.
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)

# Delete the second hypothesis
update_graph(g, c(TRUE, FALSE, TRUE, TRUE))
# Equivalent
# update_graph(g, c(1, 0, 1, 1))

}
