% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_graph.R
\name{update_graph}
\alias{update_graph}
\title{Delete multiple hypotheses from a graph}
\usage{
update_graph(graph, keep)
}
\arguments{
\item{graph}{An MCP graph as created by \code{create_graph()}}

\item{keep}{A vector coercible to Boolean, denoting which hypotheses to
delete. An entry of \code{FALSE} (or 0) corresponds to a deletion, and \code{TRUE}
(or any number other than 0) corresponds to keeping a hypothesis. We
recommend using only either a 0/1 vector or a TRUE/FALSE vector for clarity}
}
\value{
An object of class \code{updated_graph} with 3 elements
\itemize{
\item The initial graph object
\item The Boolean vector of keep/delete indicators
\item The updated graph object with appropriate hypotheses removed
}
}
\description{
Delete multiple hypotheses from a graph
}
\examples{
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)

# Delete the third hypothesis
update_graph(g, c(TRUE, TRUE, FALSE, TRUE))

}
